avivo-bot/requirements.txt
# Telegram Bot
python-telegram-bot==20.8

# Redis (async)
redis[asyncio]==5.0.3

# RAG / Embeddings
sentence-transformers==2.7.0
chromadb==0.5.0

# Local Vision (BLIP — runs on CPU, no Ollama needed)
transformers>=4.40.0
torch>=2.0.0
torchvision>=0.15.0

# LLM / Vision (via Ollama HTTP)
# httpx version is constrained by python-telegram-bot~=0.26.x, let pip resolve it
httpx>=0.26.0,<0.28.0

# FastAPI Sidecar
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Gradio Frontend
gradio==4.36.1

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
numpy==1.26.4
pillow==10.3.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0
